requirements.txt
pyautogen==0.2.25
openai>=1.3,<2.0
python-dotenv==1.0.0